+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.1 - 2020/05/01
+- Added pinlight, vividlight and exclusion
+- Fix hardlight, negation
+
 ## 2020 - 2020/05/01
 - First release